move object graph traversal into separate file
