+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/ObjectTraverser.kt
@@ -21,9 +21,7 @@
 package org.jetbrains.kotlinx.lincheck
 
 import org.jetbrains.kotlinx.lincheck.strategy.managed.ObjectLabelFactory.getObjectNumber
-import org.jetbrains.kotlinx.lincheck.util.isAtomic
-import org.jetbrains.kotlinx.lincheck.util.isAtomicFU
-import org.jetbrains.kotlinx.lincheck.util.readFieldViaUnsafe
+import org.jetbrains.kotlinx.lincheck.util.*
 import java.lang.reflect.Modifier
 import java.math.BigDecimal
 import java.math.BigInteger
