+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/util/UnsafeHolder.kt
@@ -10,7 +10,6 @@
 
 package org.jetbrains.kotlinx.lincheck.util
 
-import org.jetbrains.kotlinx.lincheck.getArrayElementOffset
 import sun.misc.Unsafe
 import java.lang.reflect.Field
 import java.lang.reflect.Modifier
